requirements.txt
flask
openai
groq
chromadb
sentence-transformers
deep-translator
twilio
python-dotenv
requests
gunicorn